@@ -1,25 +1,28 @@
 import { Component } from '@angular/core';
 
 @Component({
-    selector: 'pm-app',
-    template: `
-    <div>
-        <nav class='navbar navbar-default'>
-            <div class='container-fluid'>
-                <a class='navbar-brand'>{{pageTitle}}</a>
-                <ul class='nav navbar-nav'>
-                    <li><a [routerLink]="['/welcome']">Home</a></li>
-                    <li><a [routerLink]="['/products']">Product List</a></li>
-                    <li><a [routerLink]="['/productEdit/0']">Add Product</a></li>
-                </ul>
-            </div>
-        </nav>
-        <div class='container'>
-            <router-outlet></router-outlet>
-        </div>
-     </div>
-     `
+  selector: 'pm-root',
+  template: `
+    <nav class='navbar navbar-expand navbar-light bg-light'>
+        <a class='navbar-brand'>{{pageTitle}}</a>
+        <ul class='nav nav-pills'>
+          <li><a class='nav-link' routerLinkActive='active'
+                 [routerLink]="['/welcome']">Home</a>
+          </li>
+          <li><a class='nav-link' routerLinkActive='active' [routerLinkActiveOptions]="{ exact: true }"
+                 [routerLink]="['/products']">Product List</a>
+          </li>
+          <li><a class='nav-link' routerLinkActive='active' [routerLinkActiveOptions]="{ exact: true }"
+                 [routerLink]="['/products', 0, 'edit']">Add Product</a>
+          </li>
+        </ul>
+    </nav>
+    <div class='container'>
+      <router-outlet></router-outlet>
+    </div>
+    `,
+  styleUrls: ['./app.component.css']
 })
 export class AppComponent {
-    pageTitle: string = 'Acme Product Management';
+  pageTitle = 'Acme Product Management';
 }
