@@ -1,8 +1,8 @@
 import { Component } from '@angular/core';
 
 @Component({
-    templateUrl: 'app/home/welcome.component.html'
+  templateUrl: './welcome.component.html'
 })
 export class WelcomeComponent {
-    public pageTitle: string = 'Welcome';
+  public pageTitle = 'Welcome';
 }
