@@ -1,48 +1,46 @@
 import { Component, OnInit, OnDestroy } from '@angular/core';
-import { Router, ActivatedRoute } from '@angular/router';
+import { ActivatedRoute, Router } from '@angular/router';
 
-import { Subscription }       from 'rxjs/Subscription';
+import { Subscription } from 'rxjs';
 
-import { IProduct } from './product';
+import { Product } from './product';
 import { ProductService } from './product.service';
 
 @Component({
-    templateUrl: 'app/products/product-detail.component.html'
+  templateUrl: './product-detail.component.html',
+  styleUrls: ['./product-detail.component.css']
 })
 export class ProductDetailComponent implements OnInit, OnDestroy {
-    pageTitle: string = 'Product Detail';
-    product: IProduct;
-    errorMessage: string;
-    private sub: Subscription;
+  pageTitle = 'Product Detail';
+  errorMessage = '';
+  product: Product | undefined;
+  private sub: Subscription;
 
-    constructor(private route: ActivatedRoute,
-                private router: Router,
-                private productService: ProductService) {
-    }
+  constructor(private route: ActivatedRoute,
+    private router: Router,
+    private productService: ProductService) {
+  }
 
-    ngOnInit(): void {
-        this.sub = this.route.params.subscribe(
-            params => {
-                let id = +params['id'];
-                this.getProduct(id);
-        });
-    }
+  ngOnInit(): void {
+    this.sub = this.route.params.subscribe(
+      params => {
+        const id = +params['id'];
+        this.getProduct(id);
+      });
+  }
 
-    ngOnDestroy() {
-        this.sub.unsubscribe();
-    }
+  ngOnDestroy() {
+    this.sub.unsubscribe();
+  }
 
-    getProduct(id: number) {
-        this.productService.getProduct(id).subscribe(
-            product => this.product = product,
-            error => this.errorMessage = <any>error);
-    }
+  getProduct(id: number) {
+    this.productService.getProduct(id).subscribe(
+      product => this.product = product,
+      error => this.errorMessage = <any>error);
+  }
 
-    onBack(): void {
-        this.router.navigate(['/products']);
-    }
+  onBack(): void {
+    this.router.navigate(['/products']);
+  }
 
-    onRatingClicked(message: string): void {
-        this.pageTitle = 'Product Detail: ' + message;
-    }
 }
