@@ -1,20 +1,19 @@
 import { NgModule } from '@angular/core';
-import { RouterModule} from '@angular/router';
+import { RouterModule } from '@angular/router';
 import { ReactiveFormsModule } from '@angular/forms';
+
+import { SharedModule } from '../shared/shared.module';
 
 // Imports for loading & configuring the in-memory web api
 import { InMemoryWebApiModule } from 'angular-in-memory-web-api';
-import { ProductData }  from './product-data';
+import { ProductData } from './product-data';
 
 import { ProductListComponent } from './product-list.component';
 import { ProductDetailComponent } from './product-detail.component';
-import { ProductDetailGuard, ProductEditGuard } from './product-guard.service';
 import { ProductEditComponent } from './product-edit.component';
+import { ProductDetailGuard } from './product-detail.guard';
+import { ProductEditGuard } from './product-edit.guard';
 
-import { ProductFilterPipe } from './product-filter.pipe';
-import { ProductService } from './product.service';
-
-import { SharedModule } from '../shared/shared.module';
 
 @NgModule({
   imports: [
@@ -23,25 +22,22 @@ import { SharedModule } from '../shared/shared.module';
     InMemoryWebApiModule.forRoot(ProductData),
     RouterModule.forChild([
       { path: 'products', component: ProductListComponent },
-      { path: 'product/:id',
+      {
+        path: 'products/:id',
         canActivate: [ ProductDetailGuard],
         component: ProductDetailComponent
       },
-      { path: 'productEdit/:id',
+      {
+        path: 'products/:id/edit',
         canDeactivate: [ ProductEditGuard ],
-        component: ProductEditComponent },
+        component: ProductEditComponent
+      },
     ])
   ],
   declarations: [
     ProductListComponent,
     ProductDetailComponent,
-    ProductEditComponent,
-    ProductFilterPipe
-  ],
-  providers: [
-    ProductService,
-    ProductDetailGuard,
-    ProductEditGuard
+    ProductEditComponent
   ]
 })
-export class ProductModule {}
+export class ProductModule { }
