@@ -1,16 +1,19 @@
-import { NgModule }  from '@angular/core';
+import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { StarComponent } from './star.component';
 import { FormsModule } from '@angular/forms';
 
-import { StarComponent } from './star.component';
-
 @NgModule({
-  imports: [ CommonModule],
-  exports : [
-    CommonModule,
-    FormsModule,
+  imports: [
+    CommonModule
+  ],
+  declarations: [
     StarComponent
   ],
-  declarations: [ StarComponent ],
+  exports: [
+    StarComponent,
+    CommonModule,
+    FormsModule
+  ]
 })
 export class SharedModule { }
