@@ -1,24 +1,21 @@
-import { Component, OnChanges, Input,
-         Output, EventEmitter } from '@angular/core';
+import { Component, OnChanges, Input, EventEmitter, Output } from '@angular/core';
 
 @Component({
-    selector: 'ai-star',
-    templateUrl: 'app/shared/star.component.html',
-    styleUrls: ['app/shared/star.component.css']
+  selector: 'pm-star',
+  templateUrl: './star.component.html',
+  styleUrls: ['./star.component.css']
 })
 export class StarComponent implements OnChanges {
-    @Input() rating: number;
-    starWidth: number;
-    @Output() ratingClicked: EventEmitter<string> =
-        new EventEmitter<string>();
+  @Input() rating = 0;
+  starWidth = 0;
+  @Output() ratingClicked: EventEmitter<string> =
+    new EventEmitter<string>();
 
-    ngOnChanges(): void {
-        // Convert x out of 5 starts
-        // to y out of 86px width
-        this.starWidth = this.rating * 86 / 5;
-    }
+  ngOnChanges(): void {
+    this.starWidth = this.rating * 75 / 5;
+  }
 
-    onClick(): void {
-        this.ratingClicked.emit(`The rating ${this.rating} was clicked!`);
-    }
+  onClick(): void {
+    this.ratingClicked.emit(`The rating ${this.rating} was clicked!`);
+  }
 }
